web/build-requirements.txt
redis
python-magic
django-rq
django-drf-filepond
django-video-encoding>=1.0, <2.0
django-sortedm2m
